+++ src/components/InstancesList/InstancesList.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {
   omit,
@@ -37,12 +37,15 @@ import {
   getCurrentFilters,
   parseFiltersToStr,
   marshalInstance,
+  omitFromArray,
 } from '../../utils';
+import { INSTANCES_ID_REPORT_TIMEOUT } from '../../constants';
 import {
   InTransitItemReport,
   InstancesIdReport,
 } from '../../reports';
 import ErrorModal from '../ErrorModal';
+
 import { buildQuery } from '../../routes/buildManifestObject';
 
 import css from './instances.css';
@@ -129,8 +132,22 @@ class InstancesView extends React.Component {
   }
 
   copyInstance = (instance) => {
+    const {
+      precedingTitles,
+      succeedingTitles,
+    } = instance;
     let copiedInstance = omit(instance, ['id', 'hrid']);
+
+    if (precedingTitles?.length) {
+      copiedInstance.precedingTitles = omitFromArray(precedingTitles, 'id');
+    }
+
+    if (succeedingTitles?.length) {
+      copiedInstance.succeedingTitles = omitFromArray(succeedingTitles, 'id');
+    }
+
     copiedInstance = set(copiedInstance, 'source', 'FOLIO');
+
     this.setState({ copiedInstance });
     this.openCreateInstance();
   }
@@ -182,17 +199,30 @@ class InstancesView extends React.Component {
         reset,
         GET,
       } = this.props.parentMutator.recordsToExportIDs;
+      let infoCalloutTimer;
 
       try {
         reset();
 
+        infoCalloutTimer = setTimeout(() => {
+          sendCallout({
+            type: 'info',
+            message: <FormattedMessage id="ui-inventory.saveInstancesUIIDS.info" />,
+          });
+        }, INSTANCES_ID_REPORT_TIMEOUT);
+
         const items = await GET();
+
+        clearTimeout(infoCalloutTimer);
+
         const report = new InstancesIdReport();
 
         if (!isEmpty(items)) {
           report.toCSV(items);
         }
       } catch (error) {
+        clearTimeout(infoCalloutTimer);
+
         sendCallout({
           type: 'error',
           message: <FormattedMessage id="ui-inventory.saveInstancesUIIDS.error" />,